+++ src/LuaEngineNtop.cpp
@@ -4242,7 +4242,7 @@ static int ntop_is_local_address(lua_State *vm) {
     /* network/CIDR */
     char *ip, *mask;
     u_int8_t nmask_bits;
-    int16_t local_network_id;
+    int32_t local_network_id;
 
     shadow[0] = '\0';
     ip = shadow, mask = &shadow[1];
@@ -4280,7 +4280,7 @@ static int ntop_is_local_address(lua_State *vm) {
 static int ntop_get_address_network(lua_State *vm) {
   char *ip;
   IpAddress ipa;
-  int16_t local_network_id = -1;
+  int32_t local_network_id = -1;
 
   if (ntop_lua_check(vm, __FUNCTION__, 1, LUA_TSTRING) != CONST_LUA_OK)
     return (ntop_lua_return_value(vm, __FUNCTION__, CONST_LUA_ERROR));
@@ -4290,7 +4290,7 @@ static int ntop_get_address_network(lua_State *vm) {
   ipa.set(ip);
   ipa.isLocalHost(&local_network_id);
 
-  lua_pushinteger(vm, (u_int16_t)local_network_id);
+  lua_pushinteger(vm, (u_int32_t)local_network_id);
 
   return (ntop_lua_return_value(vm, __FUNCTION__, CONST_LUA_OK));
 }
@@ -6160,7 +6160,7 @@ static int ntop_add_local_network(lua_State *vm) {
 /* ****************************************** */
 
 static int ntop_check_local_network_alias(lua_State *vm) {
-  u_int16_t network_id = (u_int16_t)-1;
+  u_int32_t network_id = (u_int32_t)-1;
 
   ntop->getTrace()->traceEvent(TRACE_DEBUG, "%s() called", __FUNCTION__);
 
@@ -6173,10 +6173,10 @@ static int ntop_check_local_network_alias(lua_State *vm) {
     network_id = ntop->getLocalNetworkId(local_network);
 
   } else if (lua_type(vm, 1) == LUA_TNUMBER) {
-    network_id = (u_int16_t)lua_tonumber(vm, 1);
+    network_id = (u_int32_t)lua_tonumber(vm, 1);
   }
 
-  if (network_id != (u_int16_t)-1) {
+  if (network_id != (u_int32_t)-1) {
     if (!ntop->getLocalNetworkAlias(vm, network_id)) {
       lua_pushnil(vm);
       return (ntop_lua_return_value(vm, __FUNCTION__, CONST_LUA_ERROR));
@@ -6190,7 +6190,7 @@ static int ntop_check_local_network_alias(lua_State *vm) {
 
 static int ntop_get_local_network_id(lua_State *vm) {
   char *local_network;
-  u_int16_t network_id = (u_int16_t)-1;
+  u_int32_t network_id = (u_int32_t)-1;
 
   ntop->getTrace()->traceEvent(TRACE_DEBUG, "%s() called", __FUNCTION__);
 
@@ -7142,7 +7142,7 @@ static int ntop_network_name_by_id(lua_State *vm) {
 /* ****************************************** */
 
 static int ntop_network_id_by_name(lua_State *vm) {
-  u_int16_t num_local_networks = ntop->getNumLocalNetworks();
+  u_int32_t num_local_networks = ntop->getNumLocalNetworks();
   int found_id = -1;
   char *name;
 
@@ -7152,7 +7152,7 @@ static int ntop_network_id_by_name(lua_State *vm) {
     return (ntop_lua_return_value(vm, __FUNCTION__, CONST_LUA_ERROR));
   name = (char *)lua_tostring(vm, 1);
 
-  for (u_int16_t network_id = 0; network_id < num_local_networks;
+  for (u_int32_t network_id = 0; network_id < num_local_networks;
        network_id++) {
     if (!strcmp(ntop->getLocalNetworkName(network_id), name)) {
       found_id = network_id;
@@ -7167,13 +7167,13 @@ static int ntop_network_id_by_name(lua_State *vm) {
 /* ****************************************** */
 
 static int ntop_get_networks(lua_State *vm) {
-  u_int16_t num_local_networks = ntop->getNumLocalNetworks();
+  u_int32_t num_local_networks = ntop->getNumLocalNetworks();
 
   ntop->getTrace()->traceEvent(TRACE_INFO, "%s() called", __FUNCTION__);
 
   lua_newtable(vm);
 
-  for (u_int16_t network_id = 0; network_id < num_local_networks; network_id++)
+  for (u_int32_t network_id = 0; network_id < num_local_networks; network_id++)
     lua_push_uint64_table_entry(vm, ntop->getLocalNetworkName(network_id),
                                 network_id);
